refactor(app): extract components in exercise page
