+++ packages/app/src/pages/happiness-exercise/index.tsx
@@ -6,29 +6,22 @@ import React, { useEffect, useState } from "react";
 import {
   View,
   StyleSheet,
-  Modal,
   Dimensions,
-  TouchableOpacity,
-  Keyboard,
   KeyboardAvoidingView,
   Platform,
 } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { IconSvg, IconSvgName } from "components/icon";
-import { FormattedText } from "components/formatted-text";
-import { ScrollView, TextInput } from "react-native-gesture-handler";
+import { ScrollView } from "react-native-gesture-handler";
 import { useHappiness } from "context/happiness";
-import { GaussIcon } from "components/curve-icon";
-import { Trans, useTranslation } from "react-i18next";
 import { load, SOUND_NAMES, play, release } from "./sound";
 import { trackEvent } from "utils/analytics";
 import config from "config";
-import rewardDailyImg from "../../assets/images/reward-daily.gif";
-import rewardMedalImg from "../../assets/images/connection.gif";
-import rewardCertificateImg from "../../assets/images/reward-certificate.gif";
-import Gif from "pages/happiness-training/Gif";
+import { Header } from "./header";
+import { AddIdeaInput } from "./add-idea-input";
+import { CloseButton } from "./close-button";
+import { Feedback } from "./feedback";
+import { Ideas } from "./ideas";
 
-const fullWidth = Dimensions.get("window").width;
 const fullHeight = Dimensions.get("window").height;
 const FOOTER_HEIGHT = 140;
 const ADD_IDEA_HEIGHT = 120;
@@ -155,380 +148,6 @@ const styles = StyleSheet.create({
   },
   close: { position: "absolute", left: 0, bottom: 32 },
   footer: { height: FOOTER_HEIGHT },
-});
-
-type AddIdeaInputProps = { onPress: (text: string) => void };
-const AddIdeaInput = ({ onPress }: AddIdeaInputProps) => {
-  const { t } = useTranslation();
-  const [idea, setIdea] = useState("");
-  const disabled = idea.length < 2;
-  return (
-    <View style={AddIdeaInputStyles.container}>
-      <TextInput
-        multiline
-        numberOfLines={3}
-        maxLength={200}
-        style={AddIdeaInputStyles.input}
-        returnKeyType="done"
-        onChangeText={(text) => setIdea(text)}
-        placeholder={t("happiness.exercise.ideaPlaceholder")}
-        placeholderTextColor="white"
-        value={idea}
-      />
-      <TouchableOpacity
-        disabled={disabled}
-        onPress={() => {
-          Keyboard.dismiss();
-          onPress(idea);
-          setIdea("");
-        }}
-        style={AddIdeaInputStyles.iconContainer}
-      >
-        <IconSvg
-          name="tickFill"
-          size={24}
-          color={disabled ? colors.primaryThird : colors.backgroundPrimaryThird}
-          style={AddIdeaInputStyles.icon}
-        />
-      </TouchableOpacity>
-    </View>
-  );
-};
-
-const AddIdeaInputStyles = StyleSheet.create({
-  container: {
-    alignSelf: "center",
-    flexDirection: "row",
-    marginHorizontal: 32,
-    height: ADD_IDEA_HEIGHT,
-  },
-  input: {
-    flex: 1,
-    borderWidth: 5,
-    borderColor: "white",
-    paddingHorizontal: 16,
-    borderTopLeftRadius: 30,
-    borderBottomRightRadius: 30,
-    fontFamily: "IRANYekanRDMobile",
-    fontSize: 18,
-    color: "white",
-    textAlign: "right",
-    elevation: 4,
-  },
-  iconContainer: { position: "absolute", right: 15, top: 15 },
-  icon: {
-    elevation: 8,
-    borderRadius: 24,
-    backgroundColor: colors.backgroundPrimary,
-  },
-});
-
-type IdeasProps = { categoryID: string; title: string; ideas: string[] };
-function Ideas({ title, categoryID, ideas }: IdeasProps) {
-  const [modalVisible, setModalVisible] = useState(false);
-
-  return (
-    <>
-      <Modal animationType="slide" transparent={false} visible={modalVisible}>
-        <View style={ideasStyles.modal}>
-          <View style={ideasStyles.container}>
-            <FormattedText style={ideasStyles.text}>
-              {title + " "}‍
-            </FormattedText>
-            <ScrollView style={ideasStyles.list}>
-              {ideas
-                .filter((x) => x.trim().length > 2)
-                .map((idea) => (
-                  <FormattedText
-                    key={idea}
-                    style={{ color: "white", fontSize: 22 }}
-                  >
-                    {idea}
-                  </FormattedText>
-                ))}
-            </ScrollView>
-            <View style={ideasStyles.footer}>
-              <View style={ideasStyles.closeButton}>
-                <TouchableOpacity onPress={() => setModalVisible(false)}>
-                  <GaussIcon
-                    backgroundColor={"white"}
-                    color={colors.backgroundPrimaryThird}
-                    onPress={() => setModalVisible(false)}
-                    icon="tickOutline"
-                  />
-                </TouchableOpacity>
-              </View>
-              <View style={ideasStyles.imageContainer}>
-                <IconSvg
-                  name={`happinessToolbox-${categoryID}` as IconSvgName}
-                  size={70}
-                  color={colors[4]}
-                />
-              </View>
-            </View>
-          </View>
-        </View>
-      </Modal>
-      <TouchableOpacity
-        onPress={() => {
-          setModalVisible(true);
-        }}
-        style={ideasStyles.button}
-      >
-        <IconSvg
-          name={`happinessToolbox-${categoryID}` as IconSvgName}
-          size={90}
-          color={colors.backgroundPrimaryThird}
-        />
-      </TouchableOpacity>
-    </>
-  );
-}
-const ideasStyles = StyleSheet.create({
-  modal: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: colors.primaryThird,
-  },
-  container: {
-    width: "80%",
-    borderRadius: 25,
-    minHeight: fullWidth / 2,
-    backgroundColor: colors.backgroundPrimaryThird,
-    paddingVertical: 16,
-    alignItems: "center",
-  },
-  text: { fontSize: 24, color: "white" },
-  list: {
-    alignSelf: "flex-start",
-    marginVertical: 24,
-    maxHeight: fullWidth,
-    width: "100%",
-    paddingHorizontal: 32,
-  },
-  footer: { height: 100, width: "100%" },
-  imageContainer: {
-    position: "absolute",
-    right: 10,
-    bottom: 0,
-    width: 70,
-    height: 70,
-    // borderWidth: 4,
-    justifyContent: "center",
-    alignItems: "center",
-    borderColor: "white",
-    borderRadius: 70,
-  },
-  closeButton: {
-    position: "absolute",
-    left: 0,
-    bottom: 0,
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    height: 80,
-    marginTop: 6,
-    zIndex: 1,
-  },
-  button: {
-    position: "absolute",
-    bottom: 0,
-    right: 0,
-    marginBottom: 24,
-    marginHorizontal: 25,
-  },
-});
-
-type CloseButtonProps = { onPress: () => void };
-function CloseButton({ onPress }: CloseButtonProps) {
-  return (
-    <View style={closeButtonStyles.container}>
-      <GaussIcon
-        onPress={() => onPress()}
-        icon="timesFill"
-        backgroundColor="white"
-        color={colors.secondary}
-      />
-    </View>
-  );
-}
-const closeButtonStyles = StyleSheet.create({
-  container: {
-    height: 80,
-    width: 40,
-  },
-});
-
-type FeedbackProps = {
-  modalVisible: boolean;
-  title: string;
-  handleAfterDone: () => void;
-  isCategoryDone: () => boolean;
-  isAllDone: () => boolean;
-};
-const Feedback = ({
-  title,
-  modalVisible,
-  handleAfterDone,
-  isCategoryDone,
-  isAllDone,
-}: FeedbackProps) => {
-  const sound = isAllDone()
-    ? null
-    : isCategoryDone()
-    ? "reward_category"
-    : "reward_exercise";
-
-  React.useEffect(() => {
-    if (modalVisible && sound) {
-      play(sound, { stopAndPlay: true });
-    }
-  }, [sound, modalVisible]);
-
-  const text = isAllDone()
-    ? "happiness.reward.allDone"
-    : isCategoryDone()
-    ? "happiness.reward.category"
-    : "happiness.reward.exercise";
-  const image = isAllDone()
-    ? rewardCertificateImg
-    : isCategoryDone()
-    ? rewardMedalImg
-    : rewardDailyImg;
-  const image2: IconSvgName = isAllDone()
-    ? "rewardCertificate"
-    : isCategoryDone()
-    ? "rewardMedal"
-    : "rewardDaily";
-  return (
-    <Modal
-      animationType="fade"
-      hardwareAccelerated
-      transparent={false}
-      visible={modalVisible}
-      onRequestClose={() => {
-        handleAfterDone();
-      }}
-    >
-      <View style={feedbackStyles.modal}>
-        <TouchableOpacity onPress={() => handleAfterDone()}>
-          <View style={feedbackStyles.container}>
-            <View style={feedbackStyles.imageContainer}>
-              <Gif image={image} theme="purple" />
-            </View>
-            <View
-              style={{
-                position: "absolute",
-                justifyContent: "center",
-                alignItems: "center",
-                top: -20,
-              }}
-            >
-              <IconSvg
-                name={image2}
-                size={75}
-                color={colors.backgroundSecondary}
-              />
-            </View>
-            <View style={feedbackStyles.textContainer}>
-              <FormattedText style={feedbackStyles.text}>
-                <Trans
-                  i18nKey={text}
-                  values={{ title }}
-                  components={[
-                    <FormattedText
-                      style={[
-                        feedbackStyles.text,
-                        { color: colors.greenVariant },
-                      ]}
-                    />,
-                  ]}
-                />
-              </FormattedText>
-            </View>
-          </View>
-        </TouchableOpacity>
-      </View>
-    </Modal>
-  );
-};
-
-const feedbackStyles = StyleSheet.create({
-  modal: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: colors.backgroundPrimary,
-  },
-  container: {
-    width: fullWidth - 2 * 36,
-    height: fullWidth - 2 * 36,
-    borderRadius: fullWidth - 2 * 36,
-    borderWidth: 7,
-    borderColor: colors.backgroundPrimary,
-    minHeight: fullWidth / 2,
-    backgroundColor: colors.backgroundPrimaryThird,
-    // borderWidth: 3,
-    justifyContent: "center",
-    alignItems: "center",
-    elevation: 25,
-  },
-  textContainer: {
-    flexDirection: "column",
-    top: -10,
-    width: fullWidth - 36 * 2 - 130,
-  },
-  text: {
-    fontSize: 18,
-    lineHeight: 18 * 1.8,
-    color: "white",
-    textAlign: "center",
-    // borderWidth: 2,
-    // borderColor: "green",
-  },
-  imageContainer: {
-    top: -10,
-  },
-});
-
-type HeaderProps = { title: string };
-function Header({ title }: HeaderProps) {
-  return (
-    <View style={headerStyles.container}>
-      <IconSvg
-        name="cloud"
-        size={55}
-        color={colors[9]}
-        style={headerStyles.cloud1}
-      />
-      <FormattedText style={headerStyles.title}>{title}</FormattedText>
-      <IconSvg
-        name="cloud"
-        size={20}
-        color={colors[9]}
-        style={headerStyles.cloud2}
-      />
-    </View>
-  );
-}
-const headerStyles = StyleSheet.create({
-  container: {
-    flexDirection: "row",
-    justifyContent: "center",
-    paddingVertical: 30,
-    marginTop: 30,
-  },
-  cloud1: { position: "absolute", left: -8, top: 0 },
-  title: {
-    fontSize: 30,
-    color: colors.backgroundPrimaryThird,
-  },
-  cloud2: {
-    position: "absolute",
-    right: 16,
-    bottom: 15,
-  },
 });
 
 const markdownStyles = {
